Moved aws functions into a utils file Consolidated decryptFromStorage into the decrypt HH task
